Add stop hook.
This uninstalls the kafka snap during the stop hook. Stopping the
service and removing the software ensures that no processes are left
running that could hold on to a mounted volume, if storage were
attached. This also ensures that the software is cleaned up in case the
charm was co-located onto a machine with other workloads.